requirements.txt
pydantic
uvicorn
scrapy 
scrapy-playwright
pandas
scikit-learn
matplotlib
torch
protobuf
datasets
torch
transformers
fastapi[standard]
'''ipykernel'''
streamlit
python-telegram-bot
aiohttp
uvicorn
huggingface_hub
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0     


